Add files via upload
This script was built with the assistance of chatGPT. 

Greatest achievement:
At first, it was working well until suddenly the output would be a blank file. chatGPT suggested multiple workarounds with no success. I realized the root cause was that the current files were a different version of Excel so I needed to change the code to include the other extension: if file.endswith((".xlsx", ".xls")
It was also needed to install xlrd library: pip install xlrd (chatGPT suggestion)

What to improve:
if the original excel files have more than one sheet, this script will only consider the first one